packages/gcp_phone_weather/requirements.txt
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.6.2 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.16.0 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
chromedriver-binary==127.0.6516.0.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
flask==2.3.3 ; python_version >= "3.12" and python_version < "4.0"
gcp-pal==1.0.29 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.19.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.4.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.2.5 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-firestore==2.16.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.10.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.63.1 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.63.1 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.62.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.64.1 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
outcome==1.3.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
selenium==4.21.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
trio-websocket==0.11.1 ; python_version >= "3.12" and python_version < "4.0"
trio==0.25.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.1 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3[socks]==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
werkzeug==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
